refactor(server): move auth together into user domain
