+++ server/internal/user/dtos.go
@@ -2,17 +2,17 @@ package user
 
 import "github.com/gofrs/uuid/v5"
 
-type CreateDTO struct {
-	Username     string
-	PasswordHash []byte
+type userCreateDTO struct {
+	username     string
+	passwordHash []byte
 }
 
-type findOneDTO struct {
+type userFindOneDTO struct {
 	id uuid.UUID
 }
 
-type FindOneByUsernameDTO struct {
-	Username string
+type userFindOneByUsernameDTO struct {
+	username string
 }
 
 type updateDTO struct {
